+++ src/bract/core/inducer.clj
@@ -44,7 +44,7 @@
   "Run the inducers specified in the context."
   [context]
   (->> (config/ctx-inducers context)
-    (util/induce context config/apply-inducer)))
+    (util/induce context (partial config/apply-inducer-by-key (key config/ctx-inducers)))))
 
 
 (defn run-config-inducers
@@ -52,7 +52,7 @@
   [context]
   (->> (config/ctx-config context)
     config/cfg-inducers
-    (util/induce context config/apply-inducer-by-name)))
+    (util/induce context (partial config/apply-inducer-by-key (key config/cfg-inducers)))))
 
 
 (defn context-hook